refactor: sign Entity Core -> Auth Interface로 이동
