--- Projects/Core/Core/Sources/Network/APIs/SignAPI/Entity/SignInfo.swift
@@ -1,7 +1,0 @@
-public struct SignInfo {
-  public let accessToken: String
-  public let refreshToken: String
-  public let loginSuccess: Bool
-  public let schoolInfoExist: Bool
-  public let schoolInfoProvided: Bool
-}
