+++ Projects/Core/Core/Sources/Network/APIs/SignAPI/ResponseDTO/SignResponseDTO.swift
@@ -1,3 +1,5 @@
+import AuthInterface
+
 public struct SignResponseDTO: Responsable {
   public let accessToken: String?
   public let refreshToken: String?
