+++ Projects/Domain/Auth/Interface/AutoSignUseCaseInterface.swift
@@ -1,5 +1,3 @@
-import Core
-
 public protocol AutoSignUseCaseInterface {
   func execute() async throws -> SignInfo
 }
