+++ Projects/Domain/Auth/Interface/DeleteUserUseCaseInterface.swift
@@ -1,5 +1,3 @@
-import Core
-
 public protocol DeleteUserUseCaseInterface {
   func execute() async throws
 }
