+++ Projects/Domain/Auth/Interface/Entity/LoginType.swift
