+++ Projects/Domain/Auth/Interface/LogoutUseCaseInterface.swift
@@ -1,5 +1,3 @@
-import Core
-
 public protocol LogoutUseCaseInterface {
   func execute() async throws
 }
