+++ Projects/Domain/Auth/Interface/SignUpUseCaseInterface.swift
@@ -1,5 +1,3 @@
-import Core
-
 public protocol SignUpUseCaseInterface {
   func execute(loginType: LoginType) async throws -> SignInfo
 }
