+++ Projects/Feature/Agency/Project.swift
@@ -16,7 +16,8 @@ let project = Project(
             sources: ["Sources/**"],
             dependencies: [
               .project(target: "BaseFeature", path: .relativeToRoot("Projects/Feature/Base")),
-              .project(target: "CreateAgencyInterface", path: .relativeToRoot("Projects/Feature/CreateAgency"))
+              .project(target: "CreateAgencyInterface", path: .relativeToRoot("Projects/Feature/CreateAgency")),
+              .project(target: "UserInterface", path: .relativeToRoot("Projects/Domain/User"))
             ],
             settings: .settings(base: [
               "SWIFT_VERSION": "5.7"
