+++ Projects/Feature/CreateAgency/Project.swift
@@ -16,7 +16,9 @@ let project = Project(
             sources: ["Sources/**"],
             dependencies: [
               .target(name: "CreateAgencyInterface"),
-              .project(target: "BaseFeature", path: .relativeToRoot("Projects/Feature/Base"))            ]
+              .project(target: "BaseFeature", path: .relativeToRoot("Projects/Feature/Base")),
+              .project(target: "UserInterface", path: .relativeToRoot("Projects/Domain/User"))
+            ]
         ),
         Target(
           name: "CreateAgencyInterface",
