+++ Projects/Feature/CreateAgency/Sources/Scene/InputUniversityInfo/InputUniversityInfoVC.swift
@@ -5,6 +5,7 @@ import DesignSystem
 import BaseFeature
 import Core
 import CreateAgencyInterface
+import UserInterface
 
 import FlexLayout
 import PinLayout
