+++ Projects/Feature/Sign/Project.swift
@@ -25,7 +25,8 @@ let project = Project(
       sources: ["Sources/**"],
       dependencies: [
         .project(target: "BaseFeature", path: .relativeToRoot("Projects/Feature/Base")),
-        .project(target: "CreateAgencyInterface", path: .relativeToRoot("Projects/Feature/CreateAgency"))
+        .project(target: "CreateAgencyInterface", path: .relativeToRoot("Projects/Feature/CreateAgency")),
+        .project(target: "AuthInterface", path: .relativeToRoot("Projects/Domain/Auth"))
       ],
       settings: .settings(base: [
         "SWIFT_VERSION": "5.7"
